Storing location of files in a session agent
This should reduce the amount of times the file system is hit; we only
need to state where the workshop is located once, from then on we can
refer to that location through out the session.

The difference between `session` and `state` is that state can be
persisted from run to run, while session will get created when the
session is bootstrapped.