Removed the need for the start task
If the `next` task is run, and if the workshop has not been started yet,
then the functionality that used to reside in the `start` task will get
executed.

The way it detect if the workshop has been started is done sorely on
checking whether or not the *state.exs*-file exist in
the *.workshop*-folder.

This removes the necessity of documenting the `start` task.