Allow for more than one hint pr exercise
A new hint will get shown every time the user types `mix workshop.hint`
until there are no more hints to be shown.

The state is saved in the state file under :exercises and the name of
the exercise module.

The code might be a bit messy; this might be refactored later.